+++ src/index.ts
@@ -1,6 +1,7 @@
 import { ApplicationEndpoint } from "./endpoints/application";
 import { ApplicationRoleEndpoint } from "./endpoints/applicationRole";
 import { AttachmentEndpoint } from "./endpoints/attachment";
+import { AvatarEndpoint } from "./endpoints/avatar";
 import { PermissionsEndpoint } from "./endpoints/permissions";
 import { Basic, BasicAuth } from "./types";
 
@@ -49,6 +50,11 @@ export class JiraServerConnector {
     attachments: AttachmentEndpoint;
 
     /**
+    * Contains all operations from '/rest/api/latest/avatar/*'.
+    */
+    avatar: AvatarEndpoint;
+
+    /**
      * Instance new Jira Server Connector with user credentials and Jira host.
      * @param {BasicAuth} auth Basic Authorization info and Jira host 
      */
@@ -58,6 +64,7 @@ export class JiraServerConnector {
         this.applicationProperties = new ApplicationEndpoint(this.auth);
         this.applicationRoles = new ApplicationRoleEndpoint(this.auth);
         this.attachments = new AttachmentEndpoint(this.auth);
+        this.avatar = new AvatarEndpoint(this.auth);
     }